+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+* Added `BaseVectorExtensions` (spike amplitudes, locations, scalings)
+* Added `NoiseLevels` extension
 * Added `UnitLocations` extension
 * Added `templates_to_dense` method to the `utils` class for converting sparse ragged templates to dense 3D arrays. [PR #3](https://github.com/catalystneuro/ndx-spikesorting/pull/3)
 * Added `read_sorting_analyzer_from_nwb` loader script for reconstructing a SpikeInterface `SortingAnalyzer` from an ndx-spikesorting NWB file with precomputed extensions (random_spikes, templates). [PR #3](https://github.com/catalystneuro/ndx-spikesorting/pull/3)